Improve date generation
This commit ensures that data generator for the file name matches more
closely to the time that will be inserted into the report. Previously
the date and time for the name of the file was calculated at the time
the file was saved and depending on this size of the report, a
significant amount of time could pass leading to a mismatch.